Adopt suggestions made by "cargo clippy" (#72)
This is mainly writing better Rust, more than anything else.

+++ src/calculate/pass_4.rs
@@ -9,7 +9,7 @@ impl Item {
     pub fn add_attribute(&mut self, attribute_id: i32, base_value: f64, value: f64) {
         let mut attribute = Attribute::new(base_value);
         attribute.value = Some(value);
-        self.attributes.insert(attribute_id as i32, attribute);
+        self.attributes.insert(attribute_id, attribute);
     }
 }
